// frontend/src/graphQL/__generated__/SchemaDiffFragment.ts
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: SchemaDiffFragment
// ====================================================

export interface SchemaDiffFragment {
  __typename: "SchemaDiff";
  exId: string;
  schemaId: string;
  seq: number;
  uuid: string;
  zedVersion: string;
  content: any;
}